Move decodeAudioBuffer
