+++ src/player/Audio/index.ts
@@ -1,5 +1,4 @@
 export { getActiveAudioContext, autoUnlockAudio } from "./AudioContext";
-export { decodeAudioBuffer } from "./decodeAudioBuffer";
 export * from "./Sampler";
 export * from "./AudioContextEngine";
 export * from "./TimeManager";
